refactor: norm対応。球との交差判定を別ファイルに分離。
#71
