+++ Makefile
@@ -15,6 +15,7 @@ src =\
 	./draw.c \
 	./calc/rotate.c \
 	./calc/camera.c \
+	./calc/sphere.c \
 	./calc/triangle.c \
 	./move.c \
 	./loader.c \
@@ -57,6 +58,7 @@ srcs_test = \
 	./$(src_dir)/data.c \
 	./$(src_dir)/calc/rotate.c \
 	./$(src_dir)/calc/camera.c \
+	./$(src_dir)/calc/sphere.c \
 	./$(src_dir)/calc/triangle.c \
 	./$(src_dir)/debug.c \
 	./$(src_dir)/move.c \
